+++ sdk/storage/azdatalake/directory/client.go
@@ -270,7 +270,7 @@ func (d *Client) Delete(ctx context.Context, options *DeleteOptions) (DeleteResp
 	}
 }
 
-// GetProperties gets the properties of a directory.
+// GetProperties gets the properties of a directory using blob endpoint.
 func (d *Client) GetProperties(ctx context.Context, options *GetPropertiesOptions) (GetPropertiesResponse, error) {
 	opts := path.FormatGetPropertiesOptions(options)
 	var respFromCtx *http.Response
@@ -281,6 +281,14 @@ func (d *Client) GetProperties(ctx context.Context, options *GetPropertiesOption
 	}
 	newResp := path.FormatGetPropertiesResponse(&resp, respFromCtx)
 	return newResp, nil
+}
+
+// GetPathProperties gets the properties of a directory using the DFS endpoint.
+func (d *Client) GetPathProperties(ctx context.Context, options *GetPathPropertiesOptions) (GetPathPropertiesResponse, error) {
+	opts, lac, mac := path.FormatGetPathPropertiesOptions(options)
+	resp, err := d.generatedDirClientWithDFS().GetProperties(ctx, opts, lac, mac)
+	err = exported.ConvertToDFSError(err)
+	return resp, err
 }
 
 // Rename renames a directory. The original directory will no longer exist and the client will be stale.