+++ sdk/storage/azdatalake/internal/path/responses.go
@@ -18,6 +18,9 @@ type SetAccessControlResponse = generated.PathClientSetAccessControlResponse
 // GetAccessControlResponse contains the response fields for the GetAccessControl operation.
 type GetAccessControlResponse = generated.PathClientGetPropertiesResponse
 
+// GetPathPropertiesResponse contains the response fields for the GetPathProperties operation.
+type GetPathPropertiesResponse = generated.PathClientGetPropertiesResponse
+
 // UpdateAccessControlResponse contains the response fields for the UpdateAccessControlRecursive operation.
 type UpdateAccessControlResponse = generated.PathClientSetAccessControlRecursiveResponse